phase-3-ai-chatbot/backend/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
sqlmodel>=0.0.14
alembic>=1.13
psycopg2-binary>=2.9
python-dotenv>=1.0
httpx>=0.25
openai>=1.12
openai-agents>=0.1
mcp>=1.0
pytest>=7.4
pytest-asyncio>=0.23
